requirements.txt
numpy
pandas
scikit-learn
librosa
soundfile
matplotlib
streamlit
paho-mqtt
pyaudio
tqdm
joblib
tensorflow
onnx
onnx-tf
onnx2tf
torch
torchaudio
